re_str

+++ ft_strncmp.c
@@ -6,13 +6,15 @@
 /*   By: tpoungla <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/07 19:16:58 by tpoungla          #+#    #+#             */
-/*   Updated: 2022/07/21 14:45:07 by tpoungla         ###   ########.fr       */
+/*   Updated: 2022/07/22 18:39:00 by tpoungla         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strncmp(const char *s1, const char *s2, int n)
+#include <stddef.h>
+
+int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s1[i] && i < n)
